Estudiante AR finalizado

+++ src/main/java/org/backend/domain/valueobjects/Avance.java
@@ -2,15 +2,15 @@ package org.backend.domain.valueobjects;
 
 import co.com.sofka.domain.generic.ValueObject;
 
-public class Avance implements ValueObject<Long> {
-    private final Long avance;
+public class Avance implements ValueObject<Float> {
+    private final Float avance;
 
-    public Avance(Long avance) {
+    public Avance(Float avance) {
         this.avance = avance;
     }
 
     @Override
-    public Long value() {
+    public Float value() {
         return avance;
     }
 }
